Move directories around to scope withing CardsAgainstHumanity dir
